+++ src/main/java/model/PersonalAccount.java
@@ -1,12 +1,12 @@
 package model;
 
 public class PersonalAccount extends Account{
-    
+
 
     public PersonalAccount() {
     }
 
-    public PersonalAccount(long balance) {
+    public PersonalAccount(double balance) {
         super(balance);
     }
 }
